requirements.txt
asgiref==3.5.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.1
Deprecated==1.2.13
Django==4.0.4
djangorestframework==3.13.1
gcloud==0.18.3
googleapis-common-protos==1.56.0 
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
jwcrypto==1.2
oauth2client==4.1.3
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.14.1
pyparsing==3.0.8
Pyrebase4==4.5.0
python-decouple==3.6
python-jwt==3.3.2
pytz==2022.1
requests==2.27.1
requests-toolbelt==0.9.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
wrapt==1.14.0
